+++ .idea/.gitignore
@@ -1,0 +1,8 @@
+# Default ignored files
+/shelf/
+/workspace.xml
+# Datasource local storage ignored files
+/../../../../../../../:\Users\juice_pjuorme\Desktop\4IV8_ProyectoAula_Equipo4_SmartCipher\4IV8-PROYECTO_AULA-Equipo4\.idea/dataSources/
+/dataSources.local.xml
+# Editor-based HTTP Client requests
+/httpRequests/
