--- 4IV8_PROYECTO_AULA_Equipo4/ProductosServicios.html
@@ -1,44 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Productos y Servicios</title>
-    <link href="https://fonts.googleapis.com/css2?family=Fjalla+One&display=swap" rel="stylesheet">
-    <link rel="preconnect" href="https://fonts.gstatic.com">
-<link href="https://fonts.googleapis.com/css2?family=Luckiest+Guy&display=swap" rel="stylesheet">
-    <link rel="preconnect" href="https://fonts.gstatic.com">
-<link href="https://fonts.googleapis.com/css2?family=Fjalla+One&display=swap" rel="stylesheet">  
-    <link rel="stylesheet" href="CSS/estilos.css">
-    <link rel="icon" href="./img/SmartCypher.jpg">
-</head>
-<body>
-    <header class="main">
-        <h1> PRODUCTOS Y SERVICIOS</h1>
-        
-        
-    </header>
-    <div>
-        <h3>PRODUCTOS</h3>
-        <div><p>SmartCipher Es una compañia sentrada en el desarrollo de software Y
-            de aplicaciones, centradas en la solucion de problematicas actuales y 
-            que busca crear metodos accesibles para que la gente resuelva sus problemas.            
-        </p></div>
-
-        <div>
-            <h3>SOLUCIONES</h3>
-            <P> SmartCipher soluciona sus problemas, tomando datos existenes, como encuestas,
-                graficos y estudios, y los aprovecha para sacar posibles soluciones.
-
-                Tras Reunir la informacion necesaria, el equipo se pone manos a la obra, a trabajar en 
-                una variedad de lenguajes de programacion, para asi crear una página, que ayude de manera
-                simple al consumidor
-            </P>
-        </div>
-    </div>
-    <article class="main">
-        <h3> <a href="index.html">VOLVER AL IDEARIO</a></h3>
-     </article>
-</body>
-</html>
