--- 4IV8_PROYECTO_AULA_Equipo4/index.html
@@ -1,83 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=
-    <title>Ideario</title>
-    
-    <link rel="preconnect" href="https://fonts.gstatic.com">
-<link href="https://fonts.googleapis.com/css2?family=Fjalla+One&display=swap" rel="stylesheet">
-<link rel="preconnect" href="https://fonts.gstatic.com">
-<link href="https://fonts.googleapis.com/css2?family=Luckiest+Guy&display=swap" rel="stylesheet">
-    <link rel="stylesheet" href="CSS/estilos.css">
-    <link rel="preconnect" href="https://fonts.gstatic.com">
-    <link rel="icon" href="./img/SmartCypher.jpg">
-
-</head>
-<body>
-    <div>
-        <div class="main">
-            <h1>IDEARIO</h1>
-            <h3> SmartCipher Equipo 4 </h3>
-            <img src="img/SmartCypher.jpg" name="SmartCypher" class="imgS">
-        </div>
-        <div>
-            <h3>MISIÓN</h3>
-        </div>
-        <div>
-            <p>La misión de la empresa será el desarrollo de software, aplicaciones de escritorio y páginas web para la solución de problemáticas que se generen en el entorno.</p>
-        </div>
-        <div>
-            <h3>VISIÓN</h3>
-        </div>
-        <div>
-            <p>Ser una empresa de reconocido prestigio en desarrollo de software conocida por una autonomía administrativa así como con excelencia en ventas de software, respetuosa y empática con su entorno,   comprometidos con nuestros clientes y honestos en todo el proceso de desarrollo para que se brinde un producto de excelente calidad y donde el mejoramiento sea continuo en todas nuestras áreas así mismo como con nuestros consumidores.
-                Creando una creatividad con alianzas estratégicas en ámbitos nacionales e internacionales con una formación legal e integral para un buen servicio. </p>
-        </div>
-         <div>
-            <h3>POLÍTICAS DE LA EMPRESA</h3>
-        </div>
-        <div>
-            <ol>
-                <li>Todos los miembros del equipo estarán presentes durante las horas de clase</li>
-                <li>Todos los miembros del equipo ayudaran en el proceso de realizar los proyectos</li>
-                <li>Todos los miembros del equipo se trataran con respeto </li>
-                <li>Todos los miembros del equipo se presentarán a las reuniones de zoom con ropa adecuada aunque no se esté en presencial</li>
-                <li>El equipo entregarán de manera puntual todos los trabajos que se les pidan </li>
-            </ol>
-        </div>
-        <div>
-            <h3>VALORES DE LA EMPRESA</h3>
-        </div>
-        <div>
-            <ol>
-                <li>Respeto: Buscamos el respeto tanto a nuestros trabajadores, como a nuestros clientes y proveedores.</li>
-                <li>Calidad: Tenemos un fuerte compromiso a tener productos y servicios acorde a nuestras políticas.</li>
-                <li>Responsabilidad: Cumplir con los valores, políticas y trabajo en tiempo y calidad.</li>
-                <li>Compromiso: Nos dedicamos a llegar al objetivo establecido y cumplimos responsablemente con las actividades propias del puesto.</li>
-                <li>Accesibilidad:Tenemos propuesto que nuestro producto sea accesible de manera sencilla y con un precio justo a posibles clientes. </li>
-                <li>Honestidad: La Empresa será transparente para evitar malentendidos y ofrecer la imagen que merece el cliente.</li>
-                <li>Cooperación: Promovemos el trabajo en equipo como pilar fundamental para lograr las metas de la empresa.</li>
-                <li>Excelencia: Invitar a todos los integrantes de la empresa a mejorar día con día para estar siempre a la vanguardia en el mercado.</li>
-                <li>Responsabilidad Social: Crear lazos cercanos con el público y consumidores al fomentar la mejora de su bienestar y marcar su enorme efecto e importancia para el corporativo.</li>
-            </ol>
-        </div>
-        <div>
-            <h3>Filosofía</h3>
-        </div>
-        <div>
-            <p>Un trabajo con calidad,  constancia, respeto, compromiso y disciplina es el factor del éxito de nuestra empresa. Somos una empresa que nos esforzamos día a día para dar a nuestros clientes lo mejor de nosotros a fin de brindarle una buena y cordial atención y así lograr todas sus necesidades dentro del rubro en el cual nos hemos desarrollado. Estamos dispuestos a emplear todo nuestro potencial, conocimiento y calidad humana para cumplir nuestros objetivos.</p>
-        </div>
-        <article class="main">
-            <h3>CONTINUE A LAS SIGUIENTES PAGINAS</h3>
-            <hr>
-            <a href="ProductosServicios.html">PAGINA DE PRODUCTOS Y SERVICIOS</a>
-            <hr>
-            <a href="CasosExito.html">PAGINA DE CASOS DE EXITO</a>
-            <hr>
-            <a href="Contacto.html">PAGINA DE CONTACTO CON LOS DESARROLLADORES</a>
-
-        </article>
-</body>
-</html>
