+++ ProductosServicios.html
@@ -11,11 +11,11 @@
     <link rel="preconnect" href="https://fonts.gstatic.com">
 <link href="https://fonts.googleapis.com/css2?family=Fjalla+One&display=swap" rel="stylesheet">  
     <link rel="stylesheet" href="CSS/estilos.css">
+    <link rel="icon" href="./img/SmartCypher.jpg">
 </head>
 <body>
-    <header>
+    <header class="main">
         <h1> PRODUCTOS Y SERVICIOS</h1>
-        <img src="img/SmartCypher.jpg" name="SmartCypher">
         
         
     </header>
@@ -37,7 +37,7 @@
             </P>
         </div>
     </div>
-    <article>
+    <article class="main">
         <h3> <a href="index.html">VOLVER AL IDEARIO</a></h3>
      </article>
 </body>
