+++ index.html
@@ -3,7 +3,7 @@
 <head>
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <meta name="viewport" content="width=device-width, initial-scale=
     <title>Ideario</title>
     
     <link rel="preconnect" href="https://fonts.gstatic.com">
@@ -12,14 +12,15 @@
 <link href="https://fonts.googleapis.com/css2?family=Luckiest+Guy&display=swap" rel="stylesheet">
     <link rel="stylesheet" href="CSS/estilos.css">
     <link rel="preconnect" href="https://fonts.gstatic.com">
+    <link rel="icon" href="./img/SmartCypher.jpg">
 
 </head>
 <body>
     <div>
-        <div>
+        <div class="main">
             <h1>IDEARIO</h1>
-            <h3> SmartCypher Equipo 4 </h3>
-            <img src="img/SmartCypher.jpg" name="SmartCypher">
+            <h3> SmartCipher Equipo 4 </h3>
+            <img src="img/SmartCypher.jpg" name="SmartCypher" class="imgS">
         </div>
         <div>
             <h3>MISIÓN</h3>
@@ -68,7 +69,7 @@
         <div>
             <p>Un trabajo con calidad,  constancia, respeto, compromiso y disciplina es el factor del éxito de nuestra empresa. Somos una empresa que nos esforzamos día a día para dar a nuestros clientes lo mejor de nosotros a fin de brindarle una buena y cordial atención y así lograr todas sus necesidades dentro del rubro en el cual nos hemos desarrollado. Estamos dispuestos a emplear todo nuestro potencial, conocimiento y calidad humana para cumplir nuestros objetivos.</p>
         </div>
-        <article>
+        <article class="main">
             <h3>CONTINUE A LAS SIGUIENTES PAGINAS</h3>
             <hr>
             <a href="ProductosServicios.html">PAGINA DE PRODUCTOS Y SERVICIOS</a>
@@ -78,5 +79,5 @@
             <a href="Contacto.html">PAGINA DE CONTACTO CON LOS DESARROLLADORES</a>
 
         </article>
-    
+</body>
 </html>
